requirements.txt
accelerate>=0.30.0
datasets>=2.19.0
evaluate>=0.4.2
huggingface_hub>=0.23.0
numpy>=1.24.0
PyYAML>=6.0.0
torch>=2.1.0
transformers>=4.41.0
wandb>=0.17.0
